Incorrect handling of response when unable to connect to remote host.
When unable to connect to a remote host, the SOCKS server is sending only a 3
byte, rather than a 4 byte response.  The missing byte indicates the address
type (which is moot when no connection can be established).

The library now expects a 3 byte response and only reads the 4th byte if the
response indicates the command was successful.